Fixed loan creation logic and tests (#82)
* Bugfix loan requests

Loan requests didn't work properly because ModelMapper, by default,
attempts to match all that is matchable. That lead to the "accountId"
field in the CreateLoanDTO being mapped to the "id" field of the Loan
class. This then causes Hibernate to attempt to SQL update the Loan
instead of inserting it, which causes an optimistic locking exception
(or similar).

Signed-off-by: David Djuretanovic <djues3@gmail.com>

* Fixed tests

Now all tests in banking and user service pass.
The external transfer cucumber test uses mocking in the "the external
transfer is processed" step. This is due to me being unsure whether
external transfers are implemented properly.

Signed-off-by: David Djuretanovic <djues3@gmail.com>

---------

Signed-off-by: David Djuretanovic <djues3@gmail.com>